+++ src/alphabet.rs
@@ -29,6 +29,7 @@ pub(crate) enum Alphabet {
     Bengali,
     Cyrillic,
     Devanagari,
+    Ethiopic,
     Georgian,
     Greek,
     Gujarati,
@@ -37,8 +38,11 @@ pub(crate) enum Alphabet {
     Hangul,
     Hebrew,
     Hiragana,
+    Kannada,
     Katakana,
+    Lao,
     Latin,
+    Malayalam,
     Tamil,
     Telugu,
     Thai,
@@ -81,6 +85,7 @@ impl Alphabet {
             Alphabet::Bengali => &BENGALI,
             Alphabet::Cyrillic => &CYRILLIC,
             Alphabet::Devanagari => &DEVANAGARI,
+            Alphabet::Ethiopic => &ETHIOPIC,
             Alphabet::Georgian => &GEORGIAN,
             Alphabet::Greek => &GREEK,
             Alphabet::Gujarati => &GUJARATI,
@@ -89,8 +94,11 @@ impl Alphabet {
             Alphabet::Hangul => &HANGUL,
             Alphabet::Hebrew => &HEBREW,
             Alphabet::Hiragana => &HIRAGANA,
+            Alphabet::Kannada => &KANNADA,
             Alphabet::Katakana => &KATAKANA,
+            Alphabet::Lao => &LAO,
             Alphabet::Latin => &LATIN,
+            Alphabet::Malayalam => &MALAYALAM,
             Alphabet::Tamil => &TAMIL,
             Alphabet::Telugu => &TELUGU,
             Alphabet::Thai => &THAI,
@@ -141,6 +149,7 @@ static ARMENIAN: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("
 static BENGALI: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Bengali"));
 static CYRILLIC: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Cyrillic"));
 static DEVANAGARI: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Devanagari"));
+static ETHIOPIC: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Ethiopic"));
 static GEORGIAN: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Georgian"));
 static GREEK: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Greek"));
 static GUJARATI: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Gujarati"));
@@ -149,8 +158,11 @@ static HAN: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Han")
 static HANGUL: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Hangul"));
 static HEBREW: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Hebrew"));
 static HIRAGANA: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Hiragana"));
+static KANNADA: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Kannada"));
 static KATAKANA: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Katakana"));
+static LAO: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Lao"));
 static LATIN: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Latin"));
+static MALAYALAM: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Malayalam"));
 static TAMIL: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Tamil"));
 static TELUGU: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Telugu"));
 static THAI: LazyLock<CharSet> = LazyLock::new(|| CharSet::from_char_class("Thai"));
@@ -166,7 +178,7 @@ mod tests {
             Alphabet::all_supporting_single_language(),
             hashmap!(
                 Alphabet::Armenian => Language::Armenian,
-                Alphabet::Bengali => Language::Bengali,
+                Alphabet::Ethiopic => Language::Amharic,
                 Alphabet::Georgian => Language::Georgian,
                 Alphabet::Greek => Language::Greek,
                 Alphabet::Gujarati => Language::Gujarati,
@@ -174,7 +186,10 @@ mod tests {
                 Alphabet::Hangul => Language::Korean,
                 Alphabet::Hebrew => Language::Hebrew,
                 Alphabet::Hiragana => Language::Japanese,
+                Alphabet::Kannada => Language::Kannada,
                 Alphabet::Katakana => Language::Japanese,
+                Alphabet::Lao => Language::Lao,
+                Alphabet::Malayalam => Language::Malayalam,
                 Alphabet::Tamil => Language::Tamil,
                 Alphabet::Telugu => Language::Telugu,
                 Alphabet::Thai => Language::Thai