+++ tests/python/test_language.py
@@ -62,8 +62,10 @@ def test_all_languages_are_available():
         [
             Language.AFRIKAANS,
             Language.ALBANIAN,
+            Language.AMHARIC,
             Language.ARABIC,
             Language.ARMENIAN,
+            Language.ASSAMESE,
             Language.AZERBAIJANI,
             Language.BASQUE,
             Language.BELARUSIAN,
@@ -95,13 +97,17 @@ def test_all_languages_are_available():
             Language.IRISH,
             Language.ITALIAN,
             Language.JAPANESE,
+            Language.KANNADA,
             Language.KAZAKH,
             Language.KOREAN,
+            Language.KURDISH,
+            Language.LAO,
             Language.LATIN,
             Language.LATVIAN,
             Language.LITHUANIAN,
             Language.MACEDONIAN,
             Language.MALAY,
+            Language.MALAYALAM,
             Language.MAORI,
             Language.MARATHI,
             Language.MONGOLIAN,
@@ -130,6 +136,7 @@ def test_all_languages_are_available():
             Language.TURKISH,
             Language.UKRAINIAN,
             Language.URDU,
+            Language.UZBEK,
             Language.VIETNAMESE,
             Language.WELSH,
             Language.XHOSA,
@@ -144,8 +151,10 @@ def test_all_spoken_languages_are_available():
         [
             Language.AFRIKAANS,
             Language.ALBANIAN,
+            Language.AMHARIC,
             Language.ARABIC,
             Language.ARMENIAN,
+            Language.ASSAMESE,
             Language.AZERBAIJANI,
             Language.BASQUE,
             Language.BELARUSIAN,
@@ -177,12 +186,16 @@ def test_all_spoken_languages_are_available():
             Language.IRISH,
             Language.ITALIAN,
             Language.JAPANESE,
+            Language.KANNADA,
             Language.KAZAKH,
             Language.KOREAN,
+            Language.KURDISH,
+            Language.LAO,
             Language.LATVIAN,
             Language.LITHUANIAN,
             Language.MACEDONIAN,
             Language.MALAY,
+            Language.MALAYALAM,
             Language.MAORI,
             Language.MARATHI,
             Language.MONGOLIAN,
@@ -211,6 +224,7 @@ def test_all_spoken_languages_are_available():
             Language.TURKISH,
             Language.UKRAINIAN,
             Language.URDU,
+            Language.UZBEK,
             Language.VIETNAMESE,
             Language.WELSH,
             Language.XHOSA,
@@ -222,7 +236,7 @@ def test_all_spoken_languages_are_available():
 
 def test_languages_support_arabic_script():
     assert Language.all_with_arabic_script() == frozenset(
-        [Language.ARABIC, Language.PERSIAN, Language.URDU]
+        [Language.ARABIC, Language.KURDISH, Language.PERSIAN, Language.URDU]
     )
 
 
@@ -273,6 +287,7 @@ def test_languages_support_latin_script():
             Language.INDONESIAN,
             Language.IRISH,
             Language.ITALIAN,
+            Language.KURDISH,
             Language.LATIN,
             Language.LATVIAN,
             Language.LITHUANIAN,
@@ -294,6 +309,7 @@ def test_languages_support_latin_script():
             Language.TSONGA,
             Language.TSWANA,
             Language.TURKISH,
+            Language.UZBEK,
             Language.VIETNAMESE,
             Language.WELSH,
             Language.XHOSA,